Moved cache responsibility to a new file 'cache.php' and cleaned up error messages in get_google_data.php
